Rename methods file to utils in api directory, add __str__ to review & comment models
